Restructuration du projet avec maven
