--- src/com/company/Main.java
@@ -1,9 +1,0 @@
-package com.company;
-
-public class Main {
-
-    public static void main(String[] args)
-    {
-	    //launch client and server here
-    }
-}
